liberate! deny internet access permission remove weird package scanners
before you tell me i can't do this, see the current license:
3. Protecting Users' Legal Rights From Anti-Circumvention Law.

  No covered work shall be deemed part of an effective technological
measure under any applicable law fulfilling obligations under article
11 of the WIPO copyright treaty adopted on 20 December 1996, or
similar laws prohibiting or restricting circumvention of such
measures.

  When you convey a covered work, you waive any legal power to forbid
circumvention of technological measures to the extent such circumvention
is effected by exercising rights under this License with respect to
the covered work, and you disclaim any intention to limit operation or
modification of the work as a means of enforcing, against the work's
users, your or third parties' legal rights to forbid circumvention of
technological measures.

This is the corresponding Article 11

Article 11
Obligations concerning Technological Measures

Contracting Parties shall provide adequate legal protection and
effective legal remedies against the circumvention of effective
technological measures that are used by authors in connection with the
exercise of their rights under this Treaty or the Berne Convention and
that restrict acts, in respect of their works, which are not authorized
by the authors concerned or permitted by law.

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>